Added service layer and improved results page
