+++ src/main/groovy/com/frc/entity/SurveyType.groovy
@@ -1,0 +1,5 @@
+package com.frc.entity
+
+enum SurveyType {
+    PIT, MATCH
+}
