+++ src/main/resources/static/scouting.html
@@ -31,7 +31,6 @@
     <form action="" class="container" id="scouting" method="post">
 
         <div class="" id="event_info">
-
             <div class="row">
                 <div class="col-sm-6">
                     <div class="d-flex flex-column">
